+++ shedskin/cmake.py
@@ -11,20 +11,18 @@ api:
     shedskin build pkg/app.py
 
 """
+import glob
+import logging
 import modulefinder
 import os
 import pathlib
 import shutil
 import subprocess
 import sys
+import textwrap
 import time
-import logging
-import glob
 
-
-from .utils import RED, GREEN, YELLOW, RESET
-from .depend import ShedskinDependencyManager, ConanDependencyManager
-
+from .utils import CYAN, GREEN, RED, RESET, WHITE, YELLOW
 
 DEPENDENCY_GRAPH = {
     "array": [],
@@ -74,12 +72,230 @@ def pkg_path():
     sys.stdout.write(str(get_pkg_path()))
 
 
-def check_output(cmd):
-    """returns output of shell command"""
-    try:
-        return subprocess.check_output(cmd.split(), encoding="utf8").strip()
-    except FileNotFoundError:
-        return None
+class ConanDependency:
+    """ "conan dependency mixin"""
+
+    def __init__(self, name, version):
+        self.name = name
+        self.version = version
+
+    def __str__(self):
+        return f"{self.name}/{self.version}"
+
+
+class ConanBDWGC(ConanDependency):
+    """conan gc dependency"""
+
+    def __init__(
+        self,
+        name="bdwgc",
+        version="8.2.2",
+        cplusplus=True,
+        cord=False,
+        gcj_support=False,
+        java_finalization=False,
+        shared=False,
+    ):
+        self.name = name
+        self.version = version
+        self.cplusplus = cplusplus
+        self.cord = cord
+        self.gcj_support = gcj_support
+        self.java_finalization = java_finalization
+        self.shared = shared
+
+
+class ConanPCRE(ConanDependency):
+    """conan pcre dependency"""
+
+    def __init__(
+        self,
+        name="pcre",
+        version="8.45",
+        build_pcrecpp=True,
+        build_pcregrep=False,
+        shared=False,
+        with_bzip2=False,
+        with_zlib=False,
+    ):
+        self.name = name
+        self.version = version
+        self.build_pcrecpp = build_pcrecpp
+        self.build_pcregrep = build_pcregrep
+        self.shared = shared
+        self.with_bzip2 = with_bzip2
+        self.with_zlib = with_zlib
+
+
+class ConanDependencyManager:
+    """dep manager which manages and install all conan dependencies"""
+
+    def __init__(self, source_dir):
+        self.source_dir = source_dir
+        self.build_dir = self.source_dir / "build"
+        self.bdwgc = ConanBDWGC()
+        self.pcre = ConanPCRE()
+
+    def generate_conanfile(self):
+        """generate conanfile file"""
+        bdwgc = self.bdwgc
+        pcre = self.pcre
+        content = textwrap.dedent(
+            f"""
+        [requires]
+        {bdwgc}
+        {pcre}
+
+        [generators]
+        cmake_find_package
+        cmake_paths
+
+        [options]
+        bdwgc:cplusplus={bdwgc.cplusplus}
+        bdwgc:cord={bdwgc.cord}
+        bdwgc:gcj_support={bdwgc.gcj_support}
+        bdwgc:java_finalization={bdwgc.java_finalization}
+        bdwgc:shared={bdwgc.shared}
+        pcre:build_pcrecpp={pcre.build_pcrecpp}
+        pcre:build_pcregrep={pcre.build_pcregrep}
+        pcre:shared={pcre.shared}
+        pcre:with_bzip2={pcre.with_bzip2}
+        pcre:with_zlib={pcre.with_zlib}
+        """
+        )
+        conanfile = self.source_dir / "conanfile.txt"
+        if not conanfile.exists():
+            conanfile.write_text(content)
+
+    def install(self):
+        """install conan dependencies"""
+        os.system(f"cd {self.build_dir} && conan install .. --build=missing")
+
+
+class ShedskinDependencyManager:
+    """shedskin local dependency manager (SPM) class"""
+
+    def __init__(self, source_dir, reset_on_run=False):
+        self.reset_on_run = reset_on_run
+        self.source_dir = source_dir
+        self.build_dir = self.source_dir / "build"
+        self.deps_dir = self.build_dir / "deps"
+        self.include_dir = self.deps_dir / "include"
+        self.lib_dir = self.deps_dir / "lib"
+        self.downloads_dir = self.deps_dir / "downloads"
+        self.src_dir = self.deps_dir / "src"
+        self.src_dir.mkdir(parents=True, exist_ok=True)
+        self.downloads_dir.mkdir(parents=True, exist_ok=True)
+        self.lib_suffix = ".lib" if sys.platform == "win32" else ".a"
+
+        if self.reset_on_run:
+            shutil.rmtree(self.deps_dir)
+
+    def shellcmd(self, cmd, *args, **kwds):
+        """run shellcmd"""
+        print("-" * 80)
+        print(f"{WHITE}cmd{RESET}: {CYAN}{cmd}{RESET}")
+        os.system(cmd.format(*args, **kwds))
+
+    def git_clone(self, repo, to_dir):
+        """retrieve git clone of repo"""
+        self.shellcmd(f"git clone --depth=1 {repo} {to_dir}")
+
+    def cmake_generate(self, src_dir, build_dir, prefix, **options):
+        """activate cmake configuration / generation stage"""
+        opts = " ".join(f"-D{k}={v}" for k, v in options.items())
+        self.shellcmd(
+            f"cmake -S {src_dir} -B {build_dir} --install-prefix {prefix} {opts}"
+        )
+
+    def cmake_build(self, build_dir):
+        """activate cmake build stage"""
+        self.shellcmd(f"cmake --build {build_dir}")
+
+    def cmake_install(self, build_dir):
+        """activate cmake install stage"""
+        self.shellcmd(f"cmake --install {build_dir}")
+
+    def wget(self, url, output_dir):
+        """download url resource using wget"""
+        self.shellcmd(f"wget -P {output_dir} {url}")
+
+    def tar(self, archive, output_dir):
+        """uncompress tar archive"""
+        self.shellcmd(f"tar -xvf {archive} -C {output_dir}")
+
+    def targets_exist(self):
+        """check if required targets exist"""
+        libgc = self.lib_dir / f"libgc{self.lib_suffix}"
+        libgccpp = self.lib_dir / f"libgccpp{self.lib_suffix}"
+        libpcre = self.lib_dir / f"libgccpp{self.lib_suffix}"
+        gc_h = self.include_dir / "gc.h"
+        pcre_h = self.include_dir / "pcre.h"
+
+        targets = [libgc, libgccpp, libpcre, gc_h, pcre_h]
+        return all(t.exists() for t in targets)
+
+    def install_all(self):
+        """install all dependencies"""
+        if not self.targets_exist():
+            self.install_bdwgc()
+            self.install_pcre()
+        else:
+            print(f"{WHITE}SPM:{RESET} targets exist, no need to run.")
+
+    def install_bdwgc(self):
+        """download / build / install bdwgc"""
+        bdwgc_repo = "https://github.com/ivmai/bdwgc"
+        bdwgc_src = self.src_dir / "bdwgc"
+        bdwgc_build = bdwgc_src / "build"
+
+        print("download / build / install bdwgc")
+        self.git_clone(bdwgc_repo, bdwgc_src)
+        bdwgc_build.mkdir(exist_ok=True)
+        self.cmake_generate(
+            bdwgc_src,
+            bdwgc_build,
+            prefix=self.deps_dir,
+            BUILD_SHARED_LIBS=False,
+            enable_cplusplus=True,
+            build_cord=False,
+            enable_docs=False,
+            enable_gcj_support=False,
+            enable_java_finalization=False,
+        )
+        self.cmake_build(bdwgc_build)
+        self.cmake_install(bdwgc_build)
+
+    def install_pcre(self):
+        """download / build / install pcre"""
+        pcre_url = (
+            "https://sourceforge.net/projects/pcre/files/pcre/8.45/pcre-8.45.tar.gz"
+        )
+        pcre_archive = self.downloads_dir / "pcre-8.45.tar.gz"
+        pcre_src = self.src_dir / "pcre-8.45"
+        pcre_build = pcre_src / "build"
+
+        print("download / build / install pcre")
+        self.wget(pcre_url, self.downloads_dir)
+        self.tar(pcre_archive, self.src_dir)
+        # pcre_archive.unlink()
+        pcre_build.mkdir(parents=True, exist_ok=True)
+        self.cmake_generate(
+            pcre_src,
+            pcre_build,
+            prefix=self.deps_dir,
+            BUILD_SHARED_LIBS=False,
+            PCRE_BUILD_PCREGREP=False,
+            PCRE_BUILD_PCRECPP=True,
+            PCRE_SUPPORT_LIBREADLINE=False,
+            PCRE_SUPPORT_LIBEDIT=False,
+            PCRE_SUPPORT_LIBZ=False,
+            PCRE_SUPPORT_LIBBZ2=False,
+            PCRE_BUILD_TESTS=False,
+            PCRE_SHOW_REPORT=False,
+        )
+        self.cmake_build(pcre_build)
+        self.cmake_install(pcre_build)
 
 
 def add_shedskin_product(
@@ -283,13 +499,13 @@ def generate_cmakefile(gx):
         if module.builtin:
             entry = module.filename.relative_to(gx.shedskin_lib)
             entry = entry.parent / entry.stem
-            if entry.name == 'builtin': # don't include 'builtin' module
+            if entry.name == "builtin":  # don't include 'builtin' module
                 continue
             sys_mods.add(entry)
         else:
             entry = module.filename.relative_to(gx.main_module.filename.parent)
             entry = entry.parent / entry.stem
-            if entry.name == p.stem: # don't include main_module
+            if entry.name == p.stem:  # don't include main_module
                 continue
             app_mods.add(entry)
 
@@ -301,7 +517,6 @@ def generate_cmakefile(gx):
         subdir=p.parent.name,
     )
     master_clfile.write_text(master_clfile_content)
-
 
 
 class CMakeBuilder:
@@ -316,7 +531,7 @@ class CMakeBuilder:
 
     def check(self, path):
         """check file for syntax errors"""
-        with open(path, encoding='utf8') as fopen:
+        with open(path, encoding="utf8") as fopen:
             src = fopen.read()
         compile(src, path, "exec")
 
@@ -347,10 +562,12 @@ class CMakeBuilder:
                 output = subprocess.run(
                     cmd, encoding="utf-8", capture_output=True, text=True
                 ).stdout
-                assert not [l for l in output if "Traceback" in l]
+                assert not [line for line in output if "Traceback" in line]
                 for check in checks:
                     print(check)
-                    assert [l for l in output.splitlines() if l.startswith(check)]
+                    assert [
+                        line for line in output.splitlines() if line.startswith(check)
+                    ]
                 print(f"*** {GREEN}SUCCESS{RESET}:", test)
             except AssertionError:
                 print(f"*** {RED}FAILURE{RESET}:", test)
@@ -383,17 +600,19 @@ class CMakeBuilder:
     def cmake_test(self, options):
         options = " ".join(options)
         tst_cmd = f"ctest --output-on-failure {options} --test-dir {self.build_dir}"
-        print('tst_cmd:', tst_cmd)
+        print("tst_cmd:", tst_cmd)
         os.system(tst_cmd)
 
     def run_tests(self):
+        """run tests as a test runner"""
         self.process(run_tests=True)
 
     def build(self):
+        """build as a builder"""
         self.process(run_tests=False)
 
     def process(self, run_tests=False):
-        """build shedskin program"""
+        """process shedskin program with cmake"""
         start_time = time.time()
 
         cfg_options = []
@@ -464,32 +683,32 @@ class CMakeBuilder:
         # test options
 
         if run_tests:
-
             if self.options.include:
                 self.tst_options.append(f"--tests-regex {self.options.include}")
 
             if self.options.check:
-                self.check(self.options.name) # check python syntax
+                self.check(self.options.name)  # check python syntax
 
             if self.options.modified:
                 most_recent_test = pathlib.Path(self.get_most_recent_test()).stem
                 bld_options.append(f"--target {most_recent_test}")
                 tst_options.append(f"--tests-regex {most_recent_test}")
 
-            #nocleanup
+            # nocleanup
 
             if self.options.pytest:
                 try:
                     import pytest
-                    os.system('pytest')
+
+                    os.system("pytest")
                 except ImportError:
-                    print('pytest not found')
+                    print("pytest not found")
                 print()
 
             if self.options.run:
-                target_suffix = '-exe'
+                target_suffix = "-exe"
                 if self.options.extmod:
-                    target_suffix = '-ext'
+                    target_suffix = "-ext"
                 bld_options.append(f"--target {self.options.run}{target_suffix}")
                 tst_options.append(f"--tests-regex {self.options.run}")
 
@@ -510,9 +729,9 @@ class CMakeBuilder:
             if self.options.run_errs:
                 failures = self.error_tests()
                 if not failures:
-                    print(f'==> {GREEN}NO FAILURES, yay!{RESET}')
+                    print(f"==> {GREEN}NO FAILURES, yay!{RESET}")
                 else:
-                    print(f'==> {RED}TESTS FAILED:{RESET}', len(failures))
+                    print(f"==> {RED}TESTS FAILED:{RESET}", len(failures))
                     print(failures)
                     sys.exit()
 
@@ -520,3 +739,13 @@ class CMakeBuilder:
         elapsed_time = time.strftime("%H:%M:%S", time.gmtime(end_time - start_time))
         print(f"Total time: {elapsed_time}\n")
 
+
+class TestRunner(CMakeBuilder):
+    """basic test runner"""
+
+    def __init__(self, options):
+        self.options = options
+        self.build_dir = pathlib.Path("build")
+        self.source_dir = pathlib.Path.cwd()
+        self.tests = sorted(glob.glob("./test_*/test_*.py", recursive=True))
+        self.log = logging.getLogger(self.__class__.__name__)
